Crate gns_sys

Structs

CGameID

CGameID_GameID_t

CSteamID

CSteamID_SteamID_t_SteamIDComponent_t

ISteamNetworkingConnectionSignaling

ISteamNetworkingFakeUDPPort

ISteamNetworkingSignalingRecvContext

ISteamNetworkingSockets

Lower level networking API.

ISteamNetworkingSockets__bindgen_vtable

ISteamNetworkingUtils

Misc networking utilities for checking the local networking environment and estimating pings.

ISteamNetworkingUtils__bindgen_vtable

SteamDatagramGameCoordinatorServerLogin

SteamDatagramHostedAddress

SteamDatagramRelayAuthTicket

SteamIPAddress_t

SteamNetAuthenticationStatus_t

A struct used to describe our readiness to participate in authenticated, encrypted communication. In order to do this we need:

SteamNetConnectionInfo_t

Describe the state of a connection.

SteamNetConnectionRealTimeLaneStatus_t

Quick status of a particular lane

SteamNetConnectionRealTimeStatus_t

Quick connection state, pared down to something you could call more frequently without it being too big of a perf hit.

SteamNetConnectionStatusChangedCallback_t

This callback is posted whenever a connection is created, destroyed, or changes state. The m_info field will contain a complete description of the connection at the time the change occurred and the callback was posted. In particular, m_eState will have the new connection state.

SteamNetworkPingLocation_t

Object that describes a “location” on the Internet with sufficient detail that we can reasonably estimate an upper bound on the ping between the two hosts, even if a direct route between the hosts is not possible, and the connection must be routed through the Steam Datagram Relay network. This does not contain any information that identifies the host. Indeed, if two hosts are in the same building or otherwise have nearly identical networking characteristics, then it’s valid to use the same location object for both of them.

SteamNetworkingConfigValue_t

In a few places we need to set configuration options on listen sockets and connections, and have them take effect before the listen socket or connection really starts doing anything. Creating the object and then setting the options “immediately” after creation doesn’t work completely, because network packets could be received between the time the object is created and when the options are applied. To set options at creation time in a reliable way, they must be passed to the creation function. This structure is used to pass those options.

SteamNetworkingFakeIPResult_t

SteamNetworkingIPAddr

Store an IP and port. IPv6 is always used; IPv4 is represented using “IPv4-mapped” addresses: IPv4 aa.bb.cc.dd => IPv6 ::ffff:aabb:ccdd (RFC 4291 section 2.5.5.2.)

SteamNetworkingIPAddrRender

Utility class for printing a SteamNetworkingIPAddrRender.

SteamNetworkingIPAddr_IPv4MappedAddress

RFC4038, section 4.2

SteamNetworkingIdentity

An abstract way to represent the identity of a network host. All identities can be represented as simple string. Furthermore, this string representation is actually used on the wire in several places, even though it is less efficient, in order to facilitate forward compatibility. (Old client code can handle an identity type that it doesn’t understand.)

SteamNetworkingIdentityRender

Utility class for printing a SteamNetworkingIdentity. E.g. printf( “Identity is ‘%s’\n”, SteamNetworkingIdentityRender( identity ).c_str() );

SteamNetworkingMessage_t

A message that has been received.

SteamNetworkingMessagesSessionFailed_t

SteamNetworkingMessagesSessionRequest_t

SteamNetworkingPOPIDRender

Utility class for printing a SteamNetworkingPOPID.

SteamRelayNetworkStatus_t

A struct used to describe our readiness to use the relay network. To do this we first need to fetch the network configuration, which describes what POPs are available.

ValvePackingSentinel_t

__BindgenBitfieldUnit

__fsid_t

__locale_data

__locale_struct

Enums

CGameID_EGameIDType

EAccountType

EAuthSessionResponse

EBeginAuthSessionResult

EBroadcastUploadResult

EChatEntryType

EChatRoomEnterResponse

EChatSteamIDInstanceFlags

EDenyReason

EDurationControlNotification

EDurationControlOnlineState

EDurationControlProgress

EGameSearchErrorCode_t

EMarketNotAllowedReasonFlags

ENotificationPosition

EPlayerResult_t

EResult

ESteamIPType

ESteamIPv6ConnectivityProtocol

ESteamIPv6ConnectivityState

ESteamNetConnectionEnd

Enumerate various causes of connection termination. These are designed to work similar to HTTP error codes: the numeric range gives you a rough classification as to the source of the problem.

ESteamNetworkingAvailability

Describe the status of a particular network resource

ESteamNetworkingConfigDataType

ESteamNetworkingConfigScope

Configuration values can be applied to different types of objects.

ESteamNetworkingConfigValue

Configuration options

ESteamNetworkingConnectionState

High level connection status

ESteamNetworkingFakeIPType

“Fake IPs” are assigned to hosts, to make it easier to interface with older code that assumed all hosts will have an IPv4 address

ESteamNetworkingGetConfigValueResult

Return value of ISteamNetworkintgUtils::GetConfigValue

ESteamNetworkingIdentityType

Different methods of describing the identity of a network host

ESteamNetworkingSocketsDebugOutputType

Detail level for diagnostic output callback. See ISteamNetworkingUtils::SetDebugOutputFunction

EUniverse

EUserHasLicenseForAppResult

EVoiceResult

SteamNetAuthenticationStatus_t__bindgen_ty_1

SteamNetConnectionStatusChangedCallback_t__bindgen_ty_1

SteamNetworkingIPAddr__bindgen_ty_1

SteamNetworkingIdentity__bindgen_ty_1

SteamRelayNetworkStatus_t__bindgen_ty_1

_bindgen_ty_1

_bindgen_ty_2

_bindgen_ty_3

Constants

INT8_MAX

INT8_MIN

INT8_WIDTH

INT16_MAX

INT16_MIN

INT16_WIDTH

INT32_MAX

INT32_MIN

INT32_WIDTH

INT64_WIDTH

INTMAX_WIDTH

INTPTR_MAX

INTPTR_MIN

INTPTR_WIDTH

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST8_WIDTH

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST16_WIDTH

INT_FAST32_MAX

INT_FAST32_MIN

INT_FAST32_WIDTH

INT_FAST64_WIDTH

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST8_WIDTH

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST16_WIDTH

INT_LEAST32_MAX

INT_LEAST32_MIN

INT_LEAST32_WIDTH

INT_LEAST64_WIDTH

PTRDIFF_MAX

PTRDIFF_MIN

PTRDIFF_WIDTH

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIG_ATOMIC_WIDTH

SIZE_MAX

SIZE_WIDTH

STEAMNETWORKINGSOCKETS_INTERFACE_VERSION

STEAMNETWORKINGUTILS_INTERFACE_VERSION

SteamNetAuthenticationStatus_t_k_iCallback

SteamNetConnectionStatusChangedCallback_t_k_iCallback

SteamNetworkingIPAddr_k_cchMaxString

SteamNetworkingIdentity_k_cbMaxGenericBytes

SteamNetworkingIdentity_k_cchMaxGenericString

SteamNetworkingIdentity_k_cchMaxString

SteamRelayNetworkStatus_t_k_iCallback

UINT8_MAX

UINT8_WIDTH

UINT16_MAX

UINT16_WIDTH

UINT32_MAX

UINT32_WIDTH

UINT64_WIDTH

UINTMAX_WIDTH

UINTPTR_MAX

UINTPTR_WIDTH

UINT_FAST8_MAX

UINT_FAST8_WIDTH

UINT_FAST16_MAX

UINT_FAST16_WIDTH

UINT_FAST32_MAX

UINT_FAST32_WIDTH

UINT_FAST64_WIDTH

UINT_LEAST8_MAX

UINT_LEAST8_WIDTH

UINT_LEAST16_MAX

UINT_LEAST16_WIDTH

UINT_LEAST32_MAX

UINT_LEAST32_WIDTH

UINT_LEAST64_WIDTH

WCHAR_WIDTH

WINT_MAX

WINT_MIN

WINT_WIDTH

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_TYPES_LOCALE_T_H

_BITS_TYPES___LOCALE_T_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_DYNAMIC_STACK_SIZE_SOURCE

_FEATURES_H

_ISOC2X_SOURCE

_ISOC11_SOURCE

_ISOC95_SOURCE

_ISOC99_SOURCE

_LARGEFILE64_SOURCE

_LARGEFILE_SOURCE

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_STRINGS_H

_STRING_H

_SYS_CDEFS_H

_XOPEN_SOURCE

_XOPEN_SOURCE_EXTENDED

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__KERNEL_OLD_TIMEVAL_MATCHES_TIMEVAL64

__LDOUBLE_REDIRECTS_TO_FLOAT128_ABI

__OFF_T_MATCHES_OFF64_T

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_IEC_60559_BFP__

__STDC_IEC_60559_COMPLEX__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_DYNAMIC_STACK_SIZE

__USE_FORTIFY_LEVEL

__USE_GNU

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_ISOCXX11

__USE_LARGEFILE

__USE_LARGEFILE64

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_UNIX98

__USE_XOPEN

__USE_XOPEN2K

__USE_XOPEN2K8

__USE_XOPEN2K8XSI

__USE_XOPEN2KXSI

__USE_XOPEN_EXTENDED

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__glibc_c99_flexarr_available

k_HAuthTicketInvalid

k_HSteamListenSocket_Invalid

k_HSteamNetConnection_Invalid

k_HSteamNetPollGroup_Invalid

k_SteamDatagramPOPID_dev

The POPID “dev” is used in non-production environments for testing.

k_cbMaxSteamNetworkingSocketsMessageSizeSend

Max size of a single message that we can SEND. Note: We might be wiling to receive larger messages, and our peer might, too.

k_cchGameExtraInfoMax

k_cchMaxSteamNetworkingErrMsg

Max length of diagnostic error message

k_cchMaxSteamNetworkingPingLocationString

Max possible length of a ping location, in string format. This is an extremely conservative worst case value which leaves room for future syntax enhancements. Most strings in practice are a lot shorter. If you are storing many of these, you will very likely benefit from using dynamic memory.

k_cchSteamNetworkingMaxConnectionAppName

Max length of the app’s part of the description

k_cchSteamNetworkingMaxConnectionCloseReason

Max length, in bytes (including null terminator) of the reason string when a connection is closed.

k_cchSteamNetworkingMaxConnectionDescription

Max length, in bytes (include null terminator) of debug description of a connection.

k_iSteamNetworkingMessagesCallbacks

k_iSteamNetworkingSocketsCallbacks

k_iSteamNetworkingUtilsCallbacks

k_nSteamNetworkConnectionInfoFlags_DualWifi

k_nSteamNetworkConnectionInfoFlags_Fast

k_nSteamNetworkConnectionInfoFlags_LoopbackBuffers

k_nSteamNetworkConnectionInfoFlags_Relayed

k_nSteamNetworkConnectionInfoFlags_Unauthenticated

k_nSteamNetworkConnectionInfoFlags_Unencrypted

k_nSteamNetworkingConfig_P2P_Transport_ICE_Enable_All

k_nSteamNetworkingConfig_P2P_Transport_ICE_Enable_Default

k_nSteamNetworkingConfig_P2P_Transport_ICE_Enable_Disable

k_nSteamNetworkingConfig_P2P_Transport_ICE_Enable_Private

k_nSteamNetworkingConfig_P2P_Transport_ICE_Enable_Public

k_nSteamNetworkingConfig_P2P_Transport_ICE_Enable_Relay

k_nSteamNetworkingPing_Failed

Special values that are returned by some functions that return a ping.

k_nSteamNetworkingPing_Unknown

k_nSteamNetworkingSend_AutoRestartBrokenSession

k_nSteamNetworkingSend_NoDelay

k_nSteamNetworkingSend_NoNagle

k_nSteamNetworkingSend_Reliable

k_nSteamNetworkingSend_ReliableNoNagle

k_nSteamNetworkingSend_Unreliable

k_nSteamNetworkingSend_UnreliableNoDelay

k_nSteamNetworkingSend_UnreliableNoNagle

k_nSteamNetworkingSend_UseCurrentThread

k_uAPICallInvalid

k_uAppIdInvalid

k_uDepotIdInvalid

k_ulPartyBeaconIdInvalid

k_unSteamAccountIDMask

k_unSteamAccountInstanceMask

k_unSteamUserDefaultInstance

Functions

CGameID_CGameID⚠

CGameID_Render⚠

CGameID_Render1⚠

CSteamID_BValidExternalSteamID⚠

CSteamID_CSteamID⚠

CSteamID_IsValid⚠

CSteamID_Render⚠

CSteamID_Render1⚠

CSteamID_SetFromString⚠

CSteamID_SetFromStringStrict⚠

GameNetworkingSockets_Init⚠

GameNetworkingSockets_Kill⚠

ISteamNetworkingSockets_ISteamNetworkingSockets_destructor⚠

ISteamNetworkingUtils_ISteamNetworkingUtils_destructor⚠

ISteamNetworkingUtils_SetConfigValueStruct⚠

Set a configuration value, using a struct to pass the value. (This is just a convenience shortcut; see below for the implementation and a little insight into how SteamNetworkingConfigValue_t is used when setting config options during listen socket and connection creation.)

ISteamNetworkingUtils_SetConnectionConfigValueFloat⚠

ISteamNetworkingUtils_SetConnectionConfigValueInt32⚠

ISteamNetworkingUtils_SetConnectionConfigValueString⚠

ISteamNetworkingUtils_SetGlobalCallback_FakeIPResult⚠

ISteamNetworkingUtils_SetGlobalCallback_MessagesSessionFailed⚠

ISteamNetworkingUtils_SetGlobalCallback_MessagesSessionRequest⚠

ISteamNetworkingUtils_SetGlobalCallback_SteamNetAuthenticationStatusChanged⚠

ISteamNetworkingUtils_SetGlobalCallback_SteamNetConnectionStatusChanged⚠

ISteamNetworkingUtils_SetGlobalCallback_SteamRelayNetworkStatusChanged⚠

ISteamNetworkingUtils_SetGlobalConfigValueFloat⚠

ISteamNetworkingUtils_SetGlobalConfigValueInt32⚠

ISteamNetworkingUtils_SetGlobalConfigValuePtr⚠

ISteamNetworkingUtils_SetGlobalConfigValueString⚠

SteamAPI_ISteamNetworkingSockets_AcceptConnection⚠

SteamAPI_ISteamNetworkingSockets_CloseConnection⚠

SteamAPI_ISteamNetworkingSockets_CloseListenSocket⚠

SteamAPI_ISteamNetworkingSockets_ConfigureConnectionLanes⚠

SteamAPI_ISteamNetworkingSockets_ConnectByIPAddress⚠

SteamAPI_ISteamNetworkingSockets_ConnectP2P⚠

SteamAPI_ISteamNetworkingSockets_ConnectP2PCustomSignaling⚠

SteamAPI_ISteamNetworkingSockets_ConnectToHostedDedicatedServer⚠

SteamAPI_ISteamNetworkingSockets_CreateCustomSignaling⚠

Create an ISteamNetworkingConnectionSignaling object from plain C primitives.

SteamAPI_ISteamNetworkingSockets_CreateHostedDedicatedServerListenSocket⚠

SteamAPI_ISteamNetworkingSockets_CreateListenSocketIP⚠

SteamAPI_ISteamNetworkingSockets_CreateListenSocketP2P⚠

SteamAPI_ISteamNetworkingSockets_CreatePollGroup⚠

SteamAPI_ISteamNetworkingSockets_CreateSocketPair⚠

SteamAPI_ISteamNetworkingSockets_DestroyPollGroup⚠

SteamAPI_ISteamNetworkingSockets_FindRelayAuthTicketForServer⚠

SteamAPI_ISteamNetworkingSockets_FlushMessagesOnConnection⚠

SteamAPI_ISteamNetworkingSockets_GetAuthenticationStatus⚠

SteamAPI_ISteamNetworkingSockets_GetCertificateRequest⚠

SteamAPI_ISteamNetworkingSockets_GetConnectionInfo⚠

SteamAPI_ISteamNetworkingSockets_GetConnectionName⚠

SteamAPI_ISteamNetworkingSockets_GetConnectionRealTimeStatus⚠

SteamAPI_ISteamNetworkingSockets_GetConnectionUserData⚠

SteamAPI_ISteamNetworkingSockets_GetDetailedConnectionStatus⚠

SteamAPI_ISteamNetworkingSockets_GetGameCoordinatorServerLogin⚠

SteamAPI_ISteamNetworkingSockets_GetHostedDedicatedServerAddress⚠

SteamAPI_ISteamNetworkingSockets_GetHostedDedicatedServerPOPID⚠

SteamAPI_ISteamNetworkingSockets_GetHostedDedicatedServerPort⚠

SteamAPI_ISteamNetworkingSockets_GetIdentity⚠

SteamAPI_ISteamNetworkingSockets_GetListenSocketAddress⚠

SteamAPI_ISteamNetworkingSockets_InitAuthentication⚠

SteamAPI_ISteamNetworkingSockets_ReceiveMessagesOnConnection⚠

SteamAPI_ISteamNetworkingSockets_ReceiveMessagesOnPollGroup⚠

SteamAPI_ISteamNetworkingSockets_ReceivedP2PCustomSignal⚠

SteamAPI_ISteamNetworkingSockets_ReceivedP2PCustomSignal2⚠

Same as SteamAPI_ISteamNetworkingSockets_ReceivedP2PCustomSignal, but using plain C primitives.

SteamAPI_ISteamNetworkingSockets_ReceivedRelayAuthTicket⚠

SteamAPI_ISteamNetworkingSockets_RunCallbacks⚠

SteamAPI_ISteamNetworkingSockets_SendMessageToConnection⚠

SteamAPI_ISteamNetworkingSockets_SendMessages⚠

SteamAPI_ISteamNetworkingSockets_SetCertificate⚠

SteamAPI_ISteamNetworkingSockets_SetConnectionName⚠

SteamAPI_ISteamNetworkingSockets_SetConnectionPollGroup⚠

SteamAPI_ISteamNetworkingSockets_SetConnectionUserData⚠

SteamAPI_ISteamNetworkingUtils_AllocateMessage⚠

SteamAPI_ISteamNetworkingUtils_CheckPingDataUpToDate⚠

SteamAPI_ISteamNetworkingUtils_ConvertPingLocationToString⚠

SteamAPI_ISteamNetworkingUtils_EstimatePingTimeBetweenTwoLocations⚠

SteamAPI_ISteamNetworkingUtils_EstimatePingTimeFromLocalHost⚠

SteamAPI_ISteamNetworkingUtils_GetConfigValue⚠

SteamAPI_ISteamNetworkingUtils_GetConfigValueInfo⚠

SteamAPI_ISteamNetworkingUtils_GetDirectPingToPOP⚠

SteamAPI_ISteamNetworkingUtils_GetLocalPingLocation⚠

SteamAPI_ISteamNetworkingUtils_GetLocalTimestamp⚠

SteamAPI_ISteamNetworkingUtils_GetPOPCount⚠

SteamAPI_ISteamNetworkingUtils_GetPOPList⚠

SteamAPI_ISteamNetworkingUtils_GetPingToDataCenter⚠

SteamAPI_ISteamNetworkingUtils_GetRelayNetworkStatus⚠

SteamAPI_ISteamNetworkingUtils_InitRelayNetworkAccess⚠

SteamAPI_ISteamNetworkingUtils_IterateGenericEditableConfigValues⚠

SteamAPI_ISteamNetworkingUtils_ParsePingLocationString⚠

SteamAPI_ISteamNetworkingUtils_SetConfigValue⚠

SteamAPI_ISteamNetworkingUtils_SetConfigValueStruct⚠

SteamAPI_ISteamNetworkingUtils_SetConnectionConfigValueFloat⚠

SteamAPI_ISteamNetworkingUtils_SetConnectionConfigValueInt32⚠

SteamAPI_ISteamNetworkingUtils_SetConnectionConfigValueString⚠

SteamAPI_ISteamNetworkingUtils_SetDebugOutputFunction⚠

SteamAPI_ISteamNetworkingUtils_SetGlobalCallback_SteamNetAuthenticationStatusChanged⚠

SteamAPI_ISteamNetworkingUtils_SetGlobalCallback_SteamNetConnectionStatusChanged⚠

SteamAPI_ISteamNetworkingUtils_SetGlobalCallback_SteamRelayNetworkStatusChanged⚠

SteamAPI_ISteamNetworkingUtils_SetGlobalConfigValueFloat⚠

SteamAPI_ISteamNetworkingUtils_SetGlobalConfigValueInt32⚠

SteamAPI_ISteamNetworkingUtils_SetGlobalConfigValuePtr⚠

SteamAPI_ISteamNetworkingUtils_SetGlobalConfigValueString⚠

SteamAPI_SteamNetworkingIPAddr_Clear⚠

SteamAPI_SteamNetworkingIPAddr_GetIPv4⚠

SteamAPI_SteamNetworkingIPAddr_IsEqualTo⚠

SteamAPI_SteamNetworkingIPAddr_IsIPv4⚠

SteamAPI_SteamNetworkingIPAddr_IsIPv6AllZeros⚠

SteamAPI_SteamNetworkingIPAddr_IsLocalHost⚠

SteamAPI_SteamNetworkingIPAddr_ParseString⚠

SteamAPI_SteamNetworkingIPAddr_SetIPv4⚠

SteamAPI_SteamNetworkingIPAddr_SetIPv6⚠

SteamAPI_SteamNetworkingIPAddr_SetIPv6LocalHost⚠

SteamAPI_SteamNetworkingIPAddr_ToString⚠

SteamAPI_SteamNetworkingIdentity_Clear⚠

SteamAPI_SteamNetworkingIdentity_GetGenericBytes⚠

SteamAPI_SteamNetworkingIdentity_GetGenericString⚠

SteamAPI_SteamNetworkingIdentity_GetIPAddr⚠

SteamAPI_SteamNetworkingIdentity_GetSteamID⚠

SteamAPI_SteamNetworkingIdentity_GetSteamID64⚠

SteamAPI_SteamNetworkingIdentity_IsEqualTo⚠

SteamAPI_SteamNetworkingIdentity_IsInvalid⚠

SteamAPI_SteamNetworkingIdentity_IsLocalHost⚠

SteamAPI_SteamNetworkingIdentity_ParseString⚠

SteamAPI_SteamNetworkingIdentity_SetGenericBytes⚠

SteamAPI_SteamNetworkingIdentity_SetGenericString⚠

SteamAPI_SteamNetworkingIdentity_SetIPAddr⚠

SteamAPI_SteamNetworkingIdentity_SetLocalHost⚠

SteamAPI_SteamNetworkingIdentity_SetSteamID⚠

SteamAPI_SteamNetworkingIdentity_SetSteamID64⚠

SteamAPI_SteamNetworkingIdentity_ToString⚠

SteamAPI_SteamNetworkingMessage_t_Release⚠

SteamAPI_SteamNetworkingSockets_v009⚠

SteamAPI_SteamNetworkingUtils_v003⚠

SteamGameServerNetworkingSockets_LibV12⚠

SteamNetworkingIPAddr_Clear⚠

SteamNetworkingIPAddr_GetFakeIPType⚠

Classify address as FakeIP. This function never returns k_ESteamNetworkingFakeIPType_Invalid.

SteamNetworkingIPAddr_GetFakeIPType1⚠

SteamNetworkingIPAddr_GetIPv4⚠

SteamNetworkingIPAddr_IsIPv4⚠

SteamNetworkingIPAddr_IsIPv6AllZeros⚠

SteamNetworkingIPAddr_IsLocalHost⚠

SteamNetworkingIPAddr_ParseString⚠

SteamNetworkingIPAddr_SetIPv4⚠

SteamNetworkingIPAddr_SetIPv6⚠

SteamNetworkingIPAddr_SetIPv6LocalHost⚠

SteamNetworkingIPAddr_ToString⚠

SteamNetworkingIdentity_Clear⚠

SteamNetworkingIdentity_GetFakeIPType⚠

SteamNetworkingIdentity_GetGenericBytes⚠

SteamNetworkingIdentity_GetGenericString⚠

SteamNetworkingIdentity_GetIPAddr⚠

SteamNetworkingIdentity_GetIPv4⚠

SteamNetworkingIdentity_GetSteamID⚠

SteamNetworkingIdentity_GetSteamID64⚠

SteamNetworkingIdentity_IsInvalid⚠

SteamNetworkingIdentity_IsLocalHost⚠

SteamNetworkingIdentity_ParseString⚠

Parse back a string that was generated using ToString. If we don’t understand the string, but it looks “reasonable” (it matches the pattern type: and doesn’t have any funky characters, etc), then we will return true, and the type is set to k_ESteamNetworkingIdentityType_UnknownType. false will only be returned if the string looks invalid.

SteamNetworkingIdentity_ParseString1⚠

SteamNetworkingIdentity_SetGenericBytes⚠

SteamNetworkingIdentity_SetGenericString⚠

SteamNetworkingIdentity_SetIPAddr⚠

SteamNetworkingIdentity_SetIPv4Addr⚠

SteamNetworkingIdentity_SetLocalHost⚠

SteamNetworkingIdentity_SetSteamID⚠

SteamNetworkingIdentity_SetSteamID64⚠

SteamNetworkingIdentity_ToString⚠

Print to a human-readable string. This is suitable for debug messages or any other time you need to encode the identity as a string. It has a URL-like format (type:). Your buffer should be at least k_cchMaxString bytes big to avoid truncation.

SteamNetworkingIdentity_ToString1⚠

SteamNetworkingSockets_LibV12⚠

SteamNetworkingSockets_SetCustomMemoryAllocator⚠

Custom memory allocation methods. If you call this, you MUST call it exactly once, before calling any other API function. Most allocations will pass through these, especially all allocations that are per-connection. A few allocations might still go to the default CRT malloc and operator new. To use this, you must compile the library with STEAMNETWORKINGSOCKETS_ENABLE_MEM_OVERRIDE

SteamNetworkingSockets_SetLockAcquiredCallback⚠

SteamNetworkingSockets_SetLockHeldCallback⚠

SteamNetworkingSockets_SetLockWaitWarningThreshold⚠

SteamNetworkingUtils_LibV4⚠

__memcmpeq⚠

__mempcpy⚠

__stpcpy⚠

__stpncpy⚠

__strtok_r⚠

basename⚠

bcmp⚠

bcopy⚠

bzero⚠

explicit_bzero⚠

ffs⚠

ffsl⚠

ffsll⚠

index⚠

memccpy⚠

memchr⚠

memcmp⚠

memcpy⚠

memfrob⚠

memmem⚠

memmove⚠

mempcpy⚠

memrchr⚠

memset⚠

rawmemchr⚠

rindex⚠

sigabbrev_np⚠

sigdescr_np⚠

stpcpy⚠

stpncpy⚠

strcasecmp⚠

strcasecmp_l⚠

strcasestr⚠

strcat⚠

strchr⚠

strchrnul⚠

strcmp⚠

strcoll⚠

strcoll_l⚠

strcpy⚠

strcspn⚠

strdup⚠

strerror⚠

strerror_l⚠

strerror_r⚠

strerrordesc_np⚠

strerrorname_np⚠

strfry⚠

strlen⚠

strncasecmp⚠

strncasecmp_l⚠

strncat⚠

strncmp⚠

strncpy⚠

strndup⚠

strnlen⚠

strpbrk⚠

strrchr⚠

strsep⚠

strsignal⚠

strspn⚠

strstr⚠

strtok⚠

strtok_r⚠

strverscmp⚠

strxfrm⚠

strxfrm_l⚠

Type Definitions

AccountID_t

AppId_t

DepotId_t

FSteamNetworkingCustomSignalingRecvContext_OnConnectRequest

FSteamNetworkingCustomSignalingRecvContext_SendRejectionSignal

FSteamNetworkingSocketsCustomSignaling_Release

FSteamNetworkingSocketsCustomSignaling_SendSignal

FSteamNetworkingSocketsDebugOutput

Setup callback for debug output, and the desired verbosity you want.

FnSteamNetAuthenticationStatusChanged

FnSteamNetConnectionStatusChanged

FnSteamNetworkingFakeIPResult

FnSteamNetworkingMessagesSessionFailed

FnSteamNetworkingMessagesSessionRequest

FnSteamRelayNetworkStatusChanged

HAuthTicket

HSteamListenSocket

Handle used to identify a “listen socket”. Unlike traditional Berkeley sockets, a listen socket and a connection are two different abstractions.

HSteamNetConnection

Handle used to identify a connection to a remote host.

HSteamNetPollGroup

Handle used to identify a poll group, used to query many connections at once efficiently.

HSteamPipe

HSteamUser

ISteamNetworkingMessage

A message that has been received.

PFNPreMinidumpCallback

PartyBeaconID_t

RTime32

SteamAPICall_t

SteamDatagramErrMsg

SteamNetworkingErrMsg

Used to return English-language diagnostic error messages to caller. (For debugging or spewing to a console, etc. Not intended for UI.)

SteamNetworkingMicroseconds

A local timestamp. You can subtract two timestamps to get the number of elapsed microseconds. This is guaranteed to increase over time during the lifetime of a process, but not globally across runs. You don’t need to worry about the value wrapping around. Note that the underlying clock might not actually have microsecond resolution.

SteamNetworkingPOPID

Identifier used for a network location point of presence. (E.g. a Valve data center.) Typically you won’t need to directly manipulate these.

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__locale_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds64_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

compile_time_assert_type

int8

int16

int32

int64

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

intp

lint64

locale_t

size_t

uint8

uint16

uint32

uint64

uint64_steamid

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

uintp

ulint64

Unions

CGameID__bindgen_ty_1

CSteamID_SteamID_t

SteamIPAddress_t__bindgen_ty_1

SteamNetworkingConfigValue_t__bindgen_ty_1

Option value

SteamNetworkingIPAddr__bindgen_ty_2

SteamNetworkingIdentity__bindgen_ty_2