Struct gns_sys::ISteamNetworkingSockets

#[repr(C)]
pub struct ISteamNetworkingSockets {
    pub vtable_: *const ISteamNetworkingSockets__bindgen_vtable,
}

Lower level networking API.

Connection-oriented API (like TCP, not UDP). When sending and receiving messages, a connection handle is used. (For a UDP-style interface, where the peer is identified by their address with each send/recv call, see ISteamNetworkingMessages.) The typical pattern is for a “server” to “listen” on a “listen socket.” A “client” will “connect” to the server, and the server will “accept” the connection. If you have a symmetric situation where either peer may initiate the connection and server/client roles are not clearly defined, check out k_ESteamNetworkingConfig_SymmetricConnect.

But unlike TCP, it’s message-oriented, not stream-oriented.

Mix of reliable and unreliable messages

Fragmentation and reassembly

Supports connectivity over plain UDP

Also supports SDR (“Steam Datagram Relay”) connections, which are addressed by the identity of the peer. There is a “P2P” use case and a “hosted dedicated server” use case.

Note that neither of the terms “connection” nor “socket” necessarily correspond one-to-one with an underlying UDP socket. An attempt has been made to keep the semantics as similar to the standard socket model when appropriate, but some deviations do exist.

See also: ISteamNetworkingMessages, the UDP-style interface. This API might be easier to use, especially when porting existing UDP code.

Fields
vtable_: *const ISteamNetworkingSockets__bindgen_vtable
Implementations

impl ISteamNetworkingSockets

pub unsafe fn destruct(&mut self)