gns_sys

Struct SteamNetConnectionInfo_t

#[repr(C, packed(4))]
pub struct SteamNetConnectionInfo_t {

    pub m_identityRemote: SteamNetworkingIdentity,
    pub m_nUserData: int64,
    pub m_hListenSocket: HSteamListenSocket,
    pub m_addrRemote: SteamNetworkingIPAddr,
    pub m__pad1: uint16,
    pub m_idPOPRemote: SteamNetworkingPOPID,
    pub m_idPOPRelay: SteamNetworkingPOPID,
    pub m_eState: ESteamNetworkingConnectionState,
    pub m_eEndReason: c_int,
    pub m_szEndDebug: [c_char; 128],
    pub m_szConnectionDescription: [c_char; 128],
    pub m_nFlags: c_int,
    pub reserved: [uint32; 63],

}

Describe the state of a connection.

Fields
m_identityRemote: SteamNetworkingIdentity

Who is on the other end? Depending on the connection type and phase of the connection, we might not know

m_nUserData: int64

Arbitrary user data set by the local application code

m_hListenSocket: HSteamListenSocket

Handle to listen socket this was connected on, or k_HSteamListenSocket_Invalid if we initiated the connection

m_addrRemote: SteamNetworkingIPAddr

Remote address. Might be all 0โs if we donโt know it, or if this is N/A. (E.g. Basically everything except direct UDP connection.)

m__pad1: uint16m_idPOPRemote: SteamNetworkingPOPID

What data center is the remote host in? (0 if we donโt know.)

m_idPOPRelay: SteamNetworkingPOPID

What relay are we using to communicate with the remote host? (0 if not applicable.)

m_eState: ESteamNetworkingConnectionState

High level state of the connection

m_eEndReason: c_int

Basic cause of the connection termination or problem. See ESteamNetConnectionEnd for the values used

m_szEndDebug: [c_char; 128]

Human-readable, but non-localized explanation for connection termination or problem. This is intended for debugging / diagnostic purposes only, not to display to users. It might have some details specific to the issue.

m_szConnectionDescription: [c_char; 128]

Debug description. This includes the internal connection ID, connection type (and peer information), and any name given to the connection by the app. This string is used in various internal logging messages.

Note that the connection ID usually matches the HSteamNetConnection handle, but in certain cases with symmetric connections it might not.

m_nFlags: c_int

Misc flags. Bitmask of k_nSteamNetworkConnectionInfoFlags_Xxxx

reserved: [uint32; 63]

Internal stuff, room to change API easily